Factor out selector: it is not plexus specific, make it reusable.
