+++ api/maven-api-cli/src/main/java/org/apache/maven/api/cli/InvokerRequest.java
@@ -171,7 +171,9 @@ public interface InvokerRequest {
     }
 
     /**
-     * Returns a list of core extensions.
+     * Returns a list of core extensions from all sources, that were discovered and loaded. Each instance of
+     * {@link CoreExtensions} is validated, but the list elements may have overlapping elements, that requires
+     * some logic to sort out (like precedence).
      *
      * @return an {@link Optional} containing the {@link CoreExtensions}, or empty if not configured
      */
